server/requirements.txt
flask-security
flask-session
python-dotenv
redis
flask-sqlalchemy
flask-bcrypt
flask
flask_cors
requests
pandas
matplotlib
numpy
